plug-ins: Mitigate overflow in FITS import
When allocating memory for importing FITS, it was
possible for the operation to overflow the largest
datatype size, guint32. This patch adds a cast to
gsize for this operation, to reduce the risk of exceeding
the space limit before attempting to allocate.

(cherry picked from commit b01d06315352b56928625e480dd22b3876721e2c)

+++ plug-ins/file-fits/fits.c
@@ -490,11 +490,11 @@ load_image (GFile        *file,
       /* If RGB FITS image, we need to read in the whole image so we can
        * convert the planes format to RGB */
       if (hdu.naxis == 2)
-        pixels =
-          (gdouble *) g_try_malloc (width * sizeof (gdouble) * channels);
+        pixels = (gdouble *) g_try_malloc ((gsize) width * sizeof (gdouble) *
+                                           channels);
       else
-        pixels =
-          (gdouble *) g_try_malloc (width * height * sizeof (gdouble) * channels);
+        pixels = (gdouble *) g_try_malloc ((gsize) width * height *
+                                           sizeof (gdouble) * channels);
 
       if (pixels == NULL)
         {
